replace jwt verifier stub with claims provider stub
