+++ bootstrap/src/test/java/onlydust/com/marketplace/api/bootstrap/MarketplaceApiApplicationIT.java
@@ -1,9 +1,10 @@
 package onlydust.com.marketplace.api.bootstrap;
 
-import com.auth0.jwt.interfaces.JWTVerifier;
 import onlydust.com.marketplace.api.bootstrap.helper.Auth0ApiClientStub;
-import onlydust.com.marketplace.api.bootstrap.helper.JwtVerifierStub;
+import onlydust.com.marketplace.api.bootstrap.helper.Auth0ClaimsProviderStub;
 import onlydust.com.marketplace.api.domain.port.output.GithubAuthenticationPort;
+import onlydust.com.marketplace.api.rest.api.adapter.authentication.ClaimsProvider;
+import onlydust.com.marketplace.api.rest.api.adapter.authentication.auth0.Auth0JwtClaims;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
@@ -19,8 +20,8 @@ public class MarketplaceApiApplicationIT {
 
     @Bean
     @Primary
-    public JWTVerifier jwtVerifier() {
-        return new JwtVerifierStub();
+    public ClaimsProvider<Auth0JwtClaims> auth0ClaimsProvider() {
+        return new Auth0ClaimsProviderStub();
     }
 
     @Bean
