+++ bootstrap/src/test/java/onlydust/com/marketplace/api/bootstrap/it/api/ContributorSearchIT.java
@@ -1,6 +1,5 @@
 package onlydust.com.marketplace.api.bootstrap.it.api;
 
-import com.auth0.jwt.interfaces.JWTVerifier;
 import onlydust.com.marketplace.api.bootstrap.helper.UserAuthHelper;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -58,8 +57,6 @@ public class ContributorSearchIT extends AbstractMarketplaceApiIT {
             }
             """;
 
-    @Autowired
-    JWTVerifier jwtVerifier;
     @Autowired
     UserAuthHelper userAuthHelper;
